requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
openai>=1.35.0
PyPDF2>=3.0.1
python-docx>=1.1.0
pydantic>=2.7.1
tenacity>=8.2.3
